+++ src/cloudly/__init__.py
@@ -2,4 +2,4 @@
 Utilities for cloud computing.
 """
 
-__version__ = '0.3.0'
+__version__ = '0.3.2'